+++ csss/__init__.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+"""
+This is the csss Package. It contains
+csss.py and solarDisagg.py
+=================================
+This package contains the basic library modules, methods, classes and
+attributes to perform Contextually Supervise Source Separation, and the subset
+of this problem: Solar Disaggregation.
+"""
+
+from csss.CSSS import *
+import csss.SolarDisagg as SolarDisagg
